# Speed up SmoothNormals by ~2x and reduce memory allocation.